+++ s/drivers/idb/range.ts
@@ -1,9 +1,0 @@
-
-import {Scan} from "../../parts/types.js"
-
-export function idbRange({start, end}: Scan) {
-	if (start && end) return IDBKeyRange.bound(start, end)
-	if (start) return IDBKeyRange.lowerBound(start)
-	if (end) return IDBKeyRange.upperBound(end)
-}
-