+++ s/index.ts
@@ -1,15 +1,15 @@
 
-export * from "./drivers/mem.js"
-export * from "./drivers/storage.js"
-export * from "./drivers/idb.js"
-export * from "./drivers/idb/open.js"
-export * from "./drivers/idb/types.js"
+export * from "./magazines/idb.js"
+export * from "./magazines/level.js"
+export * from "./magazines/memory.js"
+export * from "./magazines/storage.js"
 
-export * from "./parts/collect.js"
-export * from "./parts/driver.js"
-export * from "./parts/store.js"
-export * from "./parts/types.js"
-export * from "./parts/writer.js"
+export * from "./utils/idb/open.js"
+export * from "./utils/idb/types.js"
+export * from "./utils/op.js"
+export * from "./utils/store-op.js"
 
 export * from "./kv.js"
+export * from "./store.js"
+export * from "./types.js"