+++ s/parts/data.ts
@@ -1,6 +1,0 @@
-
-export const Data = {
-	parse: <V>(s: string) => JSON.parse(s) as V,
-	stringify: <V>(v: V) => JSON.stringify(v),
-}
-