+++ s/volt/utils/idb/types.ts
@@ -1,6 +1,0 @@
-
-export type Idb = {
-	db: IDBDatabase
-	storeName: string
-}
-