+++ s/volt/utils/idb/wait.ts
@@ -1,8 +1,0 @@
-
-export async function idbWait<T>(request: IDBRequest) {
-	return new Promise<T>((resolve, reject) => {
-		request.onsuccess = () => resolve(request.result)
-		request.onerror = () => reject(request.error)
-	})
-}
-